Changed file name from rotatorPIDOnly to rotator
