+++ src/main/java/frc/robot/RobotContainer.java
@@ -7,7 +7,7 @@ package frc.robot;
 import frc.robot.Constants.OperatorConstants;
 import frc.robot.commands.RotatorDefaultCommand;
 import frc.robot.subsystems.Rotator;
-import frc.robot.subsystems.RotatorPIDOnly;
+import frc.robot.subsystems.Rotator;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
@@ -23,7 +23,7 @@ import edu.wpi.first.wpilibj2.command.button.Trigger;
  * subsystems, commands, and trigger mappings) should be declared here.
  */
 public class RobotContainer {
-  private final RotatorPIDOnly rotator;
+  private final Rotator rotator;
   // The robot's subsystems and commands are defined here...
 
   // Replace with CommandPS4Controller or CommandJoystick if needed
@@ -34,7 +34,7 @@ public class RobotContainer {
    * The container for the robot. Contains subsystems, OI devices, and commands.
    */
   public RobotContainer() {
-    rotator = new RotatorPIDOnly();
+    rotator = new Rotator();
     rotator.setDefaultCommand(new RotatorDefaultCommand(
         rotator,
         () -> -m_driverController.getRightX()));
